feat: Create config command
fix: English language is working again
